requirements.txt
pika
pytorch
uvicorn
fastapi
pydantic
numpy
joblib
pandas